Add command-not-found script.
This just links into the command-not-found bin that is now built. I’ve
tested it in zsh, but for some reason it doesn’t work for me in bash.
Ideally, the script could just use "alias" like:

alias command_not_found_handler = @out@/bin/command-not-found

But it doesn’t seem to work. ZSH at least needs it to be an actual function.