perf(nix-locate): improve perf of `nix-locate --hash/--package`
This commit improves performance of queries which match many entries but are
limited to a few packages. Previously, we would first process all entries and
parse them and only filter by package/hash at a very late stage. We now filter
them right when searching so we need to do less work for paths that are in the
wrong package.