default.nix: Use newer nixpkgs/Rust infrastructure
The Rust infrastructure in nixpkgs used to require explicitly
identifying a crates.io registry version to pull crates from.

Now it always fetches the current registry instead, but requires that
the sources identified by the package's Cargo.lock have a known
fixed-output hash.

So this commit updates default.nix to use a recent-enough nixpkgs (I
chose the last commit that touched Rust-related bits) and fixes the hash
for the current Cargo.lock.

Fixes #21.