fix(nix-index): don't stop on request errors
The requests stream is consumed by a `.for_each`, which stops on the
first error. This patch transforms that stream's errors into "success,
but no file list available".

Perhaps this treatment should be limited to only certain kinds of
errors, in which case this patch is over-zealous. It's pretty easy to
limit the error-filtering based on this patch, though.

Fixes #25.