Skip package sets that may not exist in older nixpkgs revisions.
When running nix-index on old nixpkgs, scopes like rPackages are missing,
so instead of exiting with an error, this change issues an error to stderr and
skips such package set.

Motivation: while it's possible to use an older version of nix-index to create
the index for older nixpkgs, newer nix-index is much more performant
so it's useful to be able to use it.